requirements.txt
matplotlib
scikit-learn
tensorflow-cpu
tensorflow-hub
numpy
pandas
opencv-python

